Data Validation

+++ Controllers/CommentController.cs
@@ -28,6 +28,11 @@ namespace FinShark.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepository.GetAllAsync();
 
             var commentDto = comments.Select(s => s.ToCommentDto());
@@ -36,12 +41,17 @@ namespace FinShark.Controllers
         }
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepository.GetByIdAsync(id);
 
-            if(comment == null)
+            if (comment == null)
             {
                 return NotFound();
             }
@@ -49,12 +59,17 @@ namespace FinShark.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stockExists = await _stockRepository.StockExists(stockId);
 
-            if(!stockExists) 
+            if (!stockExists)
             {
                 return BadRequest("Stock does not exist");
             }
@@ -63,16 +78,21 @@ namespace FinShark.Controllers
 
             await _commentRepository.CreateAsync(commentModel);
 
-            return CreatedAtAction(nameof(GetById), new {id = commentModel.Id}, commentModel.ToCommentDto());
+            return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
         [HttpPut()]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepository.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
-            if(comment == null)
+            if (comment == null)
             {
                 return NotFound("Comment not found");
             }
@@ -81,13 +101,18 @@ namespace FinShark.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
 
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var commentModel = await _commentRepository.DeleteAsync(id);
 
-            if(commentModel == null)
+            if (commentModel == null)
             {
                 return NotFound("Comment does not exist");
             }
